tf2 migration - refactor model
