+++ modelrunner.py
@@ -1,0 +1,460 @@
+import _pickle as pickle
+import os
+import time
+from rouge import FilesRouge
+
+import numpy as np
+import shutil
+import tensorflow as tf
+import tensorflow_addons as tfa
+
+import reader
+from model import Model
+from common import Common
+
+
+class ModelRunner:
+    topk = 10
+    num_batches_to_log = 100
+
+    def __init__(self, config, is_training):
+        self.config = config
+
+        self.eval_queue = None
+        self.predict_queue = None
+
+        self.eval_placeholder = None
+        self.predict_placeholder = None
+        self.eval_predicted_indices_op, self.eval_top_values_op, self.eval_true_target_strings_op, self.eval_topk_values = None, None, None, None
+        self.predict_top_indices_op, self.predict_top_scores_op, self.predict_target_strings_op = None, None, None
+        self.subtoken_to_index = None
+
+        if config.LOAD_PATH:
+            self.load_model(sess=None)
+        else:
+            with open('{}.dict.c2s'.format(config.TRAIN_PATH), 'rb') as file:
+                subtoken_to_count = pickle.load(file)
+                node_to_count = pickle.load(file)
+                target_to_count = pickle.load(file)
+                max_contexts = pickle.load(file)
+                self.num_training_examples = pickle.load(file)
+                print('Dictionaries loaded.')
+
+            if self.config.DATA_NUM_CONTEXTS <= 0:
+                self.config.DATA_NUM_CONTEXTS = max_contexts
+            self.subtoken_to_index, self.index_to_subtoken, self.subtoken_vocab_size = \
+                Common.load_vocab_from_dict(subtoken_to_count, add_values=[Common.PAD, Common.UNK],
+                                            max_size=config.SUBTOKENS_VOCAB_MAX_SIZE)
+            print('Loaded subtoken vocab. size: %d' % self.subtoken_vocab_size)
+
+            self.target_to_index, self.index_to_target, self.target_vocab_size = \
+                Common.load_vocab_from_dict(target_to_count, add_values=[Common.PAD, Common.UNK, Common.SOS],
+                                            max_size=config.TARGET_VOCAB_MAX_SIZE)
+            print('Loaded target word vocab. size: %d' % self.target_vocab_size)
+
+            self.node_to_index, self.index_to_node, self.nodes_vocab_size = \
+                Common.load_vocab_from_dict(node_to_count, add_values=[Common.PAD, Common.UNK], max_size=None)
+            print('Loaded nodes vocab. size: %d' % self.nodes_vocab_size)
+            self.epochs_trained = 0
+
+        self.dataset_reader = reader.Reader(subtoken_to_index=self.subtoken_to_index,
+                                            node_to_index=self.node_to_index,
+                                            target_to_index=self.target_to_index,
+                                            config=self.config,
+                                            is_evaluating=(not is_training))
+        batch_output = next(self.dataset_reader.get_output())
+
+        self.model = Model(batch_output, is_training)
+
+    def train(self):
+        print('Starting training')
+        start_time = time.time()
+
+        batch_num = 0
+        sum_loss = 0
+        best_f1 = 0
+        best_epoch = 0
+        best_f1_precision = 0
+        best_f1_recall = 0
+        epochs_no_improve = 0
+
+        # use the first batch to take sizes of the input data
+        optimizer, train_loss = self.build_training_graph(next(self.queue_thread.get_output()))
+        self.print_hyperparams()
+        print('Number of trainable params:',
+              np.sum([np.prod(v.get_shape().as_list()) for v in tf.trainable_variables()]))
+        self.initialize_session_variables(self.sess)
+        print('Initalized variables')
+        if self.config.LOAD_PATH:
+            self.load_model(self.sess)
+
+        time.sleep(1)
+        print('Started reader...')
+
+        multi_batch_start_time = time.time()
+        for iteration in range(1, (self.config.NUM_EPOCHS // self.config.SAVE_EVERY_EPOCHS) + 1):
+            self.queue_thread.reset(self.sess)
+            try:
+                while True:
+                    batch_num += 1
+                    _, batch_loss = self.sess.run([optimizer, train_loss])
+                    sum_loss += batch_loss
+                    # print('SINGLE BATCH LOSS', batch_loss)
+                    if batch_num % self.num_batches_to_log == 0:
+                        self.trace(sum_loss, batch_num, multi_batch_start_time)
+                        sum_loss = 0
+                        multi_batch_start_time = time.time()
+
+
+            except tf.errors.OutOfRangeError:
+                self.epochs_trained += self.config.SAVE_EVERY_EPOCHS
+                print('Finished %d epochs' % self.config.SAVE_EVERY_EPOCHS)
+                results, precision, recall, f1, rouge = self.evaluate()
+                if self.config.BEAM_WIDTH == 0:
+                    print('Accuracy after %d epochs: %.5f' % (self.epochs_trained, results))
+                else:
+                    print('Accuracy after {} epochs: {}'.format(self.epochs_trained, results))
+                print('After %d epochs: Precision: %.5f, recall: %.5f, F1: %.5f' % (
+                    self.epochs_trained, precision, recall, f1))
+                print('Rouge: ', rouge)
+                if f1 > best_f1:
+                    best_f1 = f1
+                    best_f1_precision = precision
+                    best_f1_recall = recall
+                    best_epoch = self.epochs_trained
+                    epochs_no_improve = 0
+                    self.save_model(self.sess, self.config.SAVE_PATH)
+                else:
+                    epochs_no_improve += self.config.SAVE_EVERY_EPOCHS
+                    if epochs_no_improve >= self.config.PATIENCE:
+                        print('Not improved for %d epochs, stopping training' % self.config.PATIENCE)
+                        print('Best scores - epoch %d: ' % best_epoch)
+                        print('Precision: %.5f, recall: %.5f, F1: %.5f' % (best_f1_precision, best_f1_recall, best_f1))
+                        return
+
+        if self.config.SAVE_PATH:
+            self.save_model(self.sess, self.config.SAVE_PATH + '.final')
+            print('Model saved in file: %s' % self.config.SAVE_PATH)
+
+        elapsed = int(time.time() - start_time)
+        print("Training time: %sh%sm%ss\n" % ((elapsed // 60 // 60), (elapsed // 60) % 60, elapsed % 60))
+
+    def trace(self, sum_loss, batch_num, multi_batch_start_time):
+        multi_batch_elapsed = time.time() - multi_batch_start_time
+        avg_loss = sum_loss / self.num_batches_to_log
+        print('Average loss at batch %d: %f, \tthroughput: %d samples/sec' % (batch_num, avg_loss,
+                                                                              self.config.BATCH_SIZE * self.num_batches_to_log / (
+                                                                                  multi_batch_elapsed if multi_batch_elapsed > 0 else 1)))
+
+    def evaluate(self, release=False):
+        eval_start_time = time.time()
+        if self.eval_queue is None:
+            reader_output = self.eval_queue.get_output()
+            self.eval_predicted_indices_op, self.eval_topk_values, _, _ = \
+                self.build_test_graph(reader_output)
+            self.eval_true_target_strings_op = reader_output[reader.TARGET_STRING_KEY]
+            self.saver = tf.train.Saver(max_to_keep=10)
+
+        if self.config.LOAD_PATH and not self.config.TRAIN_PATH:
+            self.initialize_session_variables(self.sess)
+            self.load_model(self.sess)
+            if release:
+                release_name = self.config.LOAD_PATH + '.release'
+                print('Releasing model, output model: %s' % release_name)
+                self.saver.save(self.sess, release_name)
+                shutil.copyfile(src=self.config.LOAD_PATH + '.dict', dst=release_name + '.dict')
+                return None
+        model_dirname = os.path.dirname(self.config.SAVE_PATH if self.config.SAVE_PATH else self.config.LOAD_PATH)
+        ref_file_name = model_dirname + '/ref.txt'
+        predicted_file_name = model_dirname + '/pred.txt'
+        if not os.path.exists(model_dirname):
+            os.makedirs(model_dirname)
+
+        with open(model_dirname + '/log.txt', 'w') as output_file, open(ref_file_name, 'w') as ref_file, open(
+                predicted_file_name,
+                'w') as pred_file:
+            num_correct_predictions = 0 if self.config.BEAM_WIDTH == 0 \
+                else np.zeros([self.config.BEAM_WIDTH], dtype=np.int32)
+            total_predictions = 0
+            total_prediction_batches = 0
+            true_positive, false_positive, false_negative = 0, 0, 0
+            self.eval_queue.reset(self.sess)
+            start_time = time.time()
+
+            try:
+                while True:
+                    predicted_indices, true_target_strings, top_values = self.sess.run(
+                        [self.eval_predicted_indices_op, self.eval_true_target_strings_op, self.eval_topk_values],
+                    )
+                    true_target_strings = Common.binary_to_string_list(true_target_strings)
+                    ref_file.write(
+                        '\n'.join(
+                            [name.replace(Common.internal_delimiter, ' ') for name in true_target_strings]) + '\n')
+                    if self.config.BEAM_WIDTH > 0:
+                        # predicted indices: (batch, time, beam_width)
+                        predicted_strings = [[[self.index_to_target[i] for i in timestep] for timestep in example] for
+                                             example in predicted_indices]
+                        predicted_strings = [list(map(list, zip(*example))) for example in
+                                             predicted_strings]  # (batch, top-k, target_length)
+                        pred_file.write('\n'.join(
+                            [' '.join(Common.filter_impossible_names(words)) for words in predicted_strings[0]]) + '\n')
+                    else:
+                        predicted_strings = [[self.index_to_target[i] for i in example]
+                                             for example in predicted_indices]
+                        pred_file.write('\n'.join(
+                            [' '.join(Common.filter_impossible_names(words)) for words in predicted_strings]) + '\n')
+
+                    num_correct_predictions = self.update_correct_predictions(num_correct_predictions, output_file,
+                                                                              zip(true_target_strings,
+                                                                                  predicted_strings))
+                    true_positive, false_positive, false_negative = self.update_per_subtoken_statistics(
+                        zip(true_target_strings, predicted_strings),
+                        true_positive, false_positive, false_negative)
+
+                    total_predictions += len(true_target_strings)
+                    total_prediction_batches += 1
+                    if total_prediction_batches % self.num_batches_to_log == 0:
+                        elapsed = time.time() - start_time
+                        self.trace_evaluation(output_file, num_correct_predictions, total_predictions, elapsed)
+            except tf.errors.OutOfRangeError:
+                pass
+
+            print('Done testing, epoch reached')
+            output_file.write(str(num_correct_predictions / total_predictions) + '\n')
+            # Common.compute_bleu(ref_file_name, predicted_file_name)
+
+        elapsed = int(time.time() - eval_start_time)
+        precision, recall, f1 = self.calculate_results(true_positive, false_positive, false_negative)
+        files_rouge = FilesRouge(predicted_file_name, ref_file_name)
+        rouge = files_rouge.get_scores(avg=True, ignore_empty=True)
+        print("Evaluation time: %sh%sm%ss" % ((elapsed // 60 // 60), (elapsed // 60) % 60, elapsed % 60))
+        return num_correct_predictions / total_predictions, \
+               precision, recall, f1, rouge
+
+    def update_correct_predictions(self, num_correct_predictions, output_file, results):
+        for original_name, predicted in results:
+            original_name_parts = original_name.split(Common.internal_delimiter)  # list
+            filtered_original = Common.filter_impossible_names(original_name_parts)  # list
+            predicted_first = predicted
+            if self.config.BEAM_WIDTH > 0:
+                predicted_first = predicted[0]
+            filtered_predicted_first_parts = Common.filter_impossible_names(predicted_first)  # list
+
+            if self.config.BEAM_WIDTH == 0:
+                output_file.write('Original: ' + Common.internal_delimiter.join(original_name_parts) +
+                                  ' , predicted 1st: ' + Common.internal_delimiter.join(
+                    filtered_predicted_first_parts) + '\n')
+                if filtered_original == filtered_predicted_first_parts or Common.unique(
+                        filtered_original) == Common.unique(
+                    filtered_predicted_first_parts) or ''.join(filtered_original) == ''.join(
+                    filtered_predicted_first_parts):
+                    num_correct_predictions += 1
+            else:
+                filtered_predicted = [Common.internal_delimiter.join(Common.filter_impossible_names(p)) for p in
+                                      predicted]
+
+                true_ref = original_name
+                output_file.write('Original: ' + ' '.join(original_name_parts) + '\n')
+                for i, p in enumerate(filtered_predicted):
+                    output_file.write('\t@{}: {}'.format(i + 1, ' '.join(p.split(Common.internal_delimiter))) + '\n')
+                if true_ref in filtered_predicted:
+                    index_of_correct = filtered_predicted.index(true_ref)
+                    update = np.concatenate(
+                        [np.zeros(index_of_correct, dtype=np.int32),
+                         np.ones(self.config.BEAM_WIDTH - index_of_correct, dtype=np.int32)])
+                    num_correct_predictions += update
+        return num_correct_predictions
+
+    def update_per_subtoken_statistics(self, results, true_positive, false_positive, false_negative):
+        for original_name, predicted in results:
+            if self.config.BEAM_WIDTH > 0:
+                predicted = predicted[0]
+            filtered_predicted_names = Common.filter_impossible_names(predicted)
+            filtered_original_subtokens = Common.filter_impossible_names(original_name.split(Common.internal_delimiter))
+
+            if ''.join(filtered_original_subtokens) == ''.join(filtered_predicted_names):
+                true_positive += len(filtered_original_subtokens)
+                continue
+
+            for subtok in filtered_predicted_names:
+                if subtok in filtered_original_subtokens:
+                    true_positive += 1
+                else:
+                    false_positive += 1
+            for subtok in filtered_original_subtokens:
+                if not subtok in filtered_predicted_names:
+                    false_negative += 1
+        return true_positive, false_positive, false_negative
+
+    def print_hyperparams(self):
+        print('Training batch size:\t\t\t', self.config.BATCH_SIZE)
+        print('Dataset path:\t\t\t\t', self.config.TRAIN_PATH)
+        print('Training file path:\t\t\t', self.config.TRAIN_PATH + '.train.c2s')
+        print('Validation path:\t\t\t', self.config.TEST_PATH)
+        print('Taking max contexts from each example:\t', self.config.MAX_CONTEXTS)
+        print('Random path sampling:\t\t\t', self.config.RANDOM_CONTEXTS)
+        print('Embedding size:\t\t\t\t', self.config.EMBEDDINGS_SIZE)
+        if self.config.BIRNN:
+            print('Using BiLSTMs, each of size:\t\t', self.config.RNN_SIZE // 2)
+        else:
+            print('Uni-directional LSTM of size:\t\t', self.config.RNN_SIZE)
+        print('Decoder size:\t\t\t\t', self.config.DECODER_SIZE)
+        print('Decoder layers:\t\t\t\t', self.config.NUM_DECODER_LAYERS)
+        print('Max path lengths:\t\t\t', self.config.MAX_PATH_LENGTH)
+        print('Max subtokens in a token:\t\t', self.config.MAX_NAME_PARTS)
+        print('Max target length:\t\t\t', self.config.MAX_TARGET_PARTS)
+        print('Embeddings dropout keep_prob:\t\t', self.config.EMBEDDINGS_DROPOUT_KEEP_PROB)
+        print('LSTM dropout keep_prob:\t\t\t', self.config.RNN_DROPOUT_KEEP_PROB)
+        print('============================================')
+
+    @staticmethod
+    def calculate_results(true_positive, false_positive, false_negative):
+        if true_positive + false_positive > 0:
+            precision = true_positive / (true_positive + false_positive)
+        else:
+            precision = 0
+        if true_positive + false_negative > 0:
+            recall = true_positive / (true_positive + false_negative)
+        else:
+            recall = 0
+        if precision + recall > 0:
+            f1 = 2 * precision * recall / (precision + recall)
+        else:
+            f1 = 0
+        return precision, recall, f1
+
+    @staticmethod
+    def trace_evaluation(output_file, correct_predictions, total_predictions, elapsed):
+        accuracy_message = str(correct_predictions / total_predictions)
+        throughput_message = "Prediction throughput: %d" % int(total_predictions / (elapsed if elapsed > 0 else 1))
+        output_file.write(accuracy_message + '\n')
+        output_file.write(throughput_message)
+        # print(accuracy_message)
+        print(throughput_message)
+
+    def predict(self, predict_data_lines):
+        if self.predict_queue is None:
+            self.predict_queue = reader.Reader(subtoken_to_index=self.subtoken_to_index,
+                                               node_to_index=self.node_to_index,
+                                               target_to_index=self.target_to_index,
+                                               config=self.config, is_evaluating=True)
+            self.predict_placeholder = tf.placeholder(tf.string)
+            reader_output = self.predict_queue.process_from_placeholder(self.predict_placeholder)
+            reader_output = {key: tf.expand_dims(tensor, 0) for key, tensor in reader_output.items()}
+            self.predict_top_indices_op, self.predict_top_scores_op, _, self.attention_weights_op = \
+                self.build_test_graph(reader_output)
+            self.predict_source_string = reader_output[reader.PATH_SOURCE_STRINGS_KEY]
+            self.predict_path_string = reader_output[reader.PATH_STRINGS_KEY]
+            self.predict_path_target_string = reader_output[reader.PATH_TARGET_STRINGS_KEY]
+            self.predict_target_strings_op = reader_output[reader.TARGET_STRING_KEY]
+
+            self.initialize_session_variables(self.sess)
+            self.saver = tf.train.Saver()
+            self.load_model(self.sess)
+
+        results = []
+        for line in predict_data_lines:
+            predicted_indices, top_scores, true_target_strings, attention_weights, path_source_string, path_strings, path_target_string = self.sess.run(
+                [self.predict_top_indices_op, self.predict_top_scores_op, self.predict_target_strings_op,
+                 self.attention_weights_op,
+                 self.predict_source_string, self.predict_path_string, self.predict_path_target_string],
+                feed_dict={self.predict_placeholder: line})
+
+            top_scores = np.squeeze(top_scores, axis=0)
+            path_source_string = path_source_string.reshape((-1))
+            path_strings = path_strings.reshape((-1))
+            path_target_string = path_target_string.reshape((-1))
+            predicted_indices = np.squeeze(predicted_indices, axis=0)
+            true_target_strings = Common.binary_to_string(true_target_strings[0])
+
+            if self.config.BEAM_WIDTH > 0:
+                predicted_strings = [[self.index_to_target[sugg] for sugg in timestep]
+                                     for timestep in predicted_indices]  # (target_length, top-k)  
+                predicted_strings = list(map(list, zip(*predicted_strings)))  # (top-k, target_length)
+                top_scores = [np.exp(np.sum(s)) for s in zip(*top_scores)]
+            else:
+                predicted_strings = [self.index_to_target[idx]
+                                     for idx in predicted_indices]  # (batch, target_length)  
+
+            attention_per_path = None
+            if self.config.BEAM_WIDTH == 0:
+                attention_per_path = self.get_attention_per_path(path_source_string, path_strings, path_target_string,
+                                                                 attention_weights)
+
+            results.append((true_target_strings, predicted_strings, top_scores, attention_per_path))
+        return results
+
+    @staticmethod
+    def get_attention_per_path(source_strings, path_strings, target_strings, attention_weights):
+        # attention_weights:  (time, contexts)
+        results = []
+        for time_step in attention_weights:
+            attention_per_context = {}
+            for source, path, target, weight in zip(source_strings, path_strings, target_strings, time_step):
+                string_triplet = (
+                    Common.binary_to_string(source), Common.binary_to_string(path), Common.binary_to_string(target))
+                attention_per_context[string_triplet] = weight
+            results.append(attention_per_context)
+        return results
+
+    def save_model(self, sess, path):
+        save_target = path + '_iter%d' % self.epochs_trained
+        dirname = os.path.dirname(save_target)
+        if not os.path.exists(dirname):
+            os.makedirs(dirname)
+        self.saver.save(sess, save_target)
+
+        dictionaries_path = save_target + '.dict'
+        with open(dictionaries_path, 'wb') as file:
+            pickle.dump(self.subtoken_to_index, file)
+            pickle.dump(self.index_to_subtoken, file)
+            pickle.dump(self.subtoken_vocab_size, file)
+
+            pickle.dump(self.target_to_index, file)
+            pickle.dump(self.index_to_target, file)
+            pickle.dump(self.target_vocab_size, file)
+
+            pickle.dump(self.node_to_index, file)
+            pickle.dump(self.index_to_node, file)
+            pickle.dump(self.nodes_vocab_size, file)
+
+            pickle.dump(self.num_training_examples, file)
+            pickle.dump(self.epochs_trained, file)
+            pickle.dump(self.config, file)
+        print('Saved after %d epochs in: %s' % (self.epochs_trained, save_target))
+
+    def load_model(self, sess):
+        if not sess is None:
+            self.saver.restore(sess, self.config.LOAD_PATH)
+            print('Done loading model')
+        with open(self.config.LOAD_PATH + '.dict', 'rb') as file:
+            if self.subtoken_to_index is not None:
+                return
+            print('Loading dictionaries from: ' + self.config.LOAD_PATH)
+            self.subtoken_to_index = pickle.load(file)
+            self.index_to_subtoken = pickle.load(file)
+            self.subtoken_vocab_size = pickle.load(file)
+
+            self.target_to_index = pickle.load(file)
+            self.index_to_target = pickle.load(file)
+            self.target_vocab_size = pickle.load(file)
+
+            self.node_to_index = pickle.load(file)
+            self.index_to_node = pickle.load(file)
+            self.nodes_vocab_size = pickle.load(file)
+
+            self.num_training_examples = pickle.load(file)
+            self.epochs_trained = pickle.load(file)
+            saved_config = pickle.load(file)
+            self.config.take_model_hyperparams_from(saved_config)
+            print('Done loading dictionaries')
+
+    @staticmethod
+    def initialize_session_variables(sess):
+        sess.run(tf.group(tf.global_variables_initializer(), tf.local_variables_initializer(), tf.tables_initializer()))
+
+    def get_should_reuse_variables(self):
+        if self.config.TRAIN_PATH:
+            return True
+        else:
+            return None
